[fix] 修正 API CRUD 操作
